pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "inference-streaming-benchmark"
dynamic = ["version"]
description = "A test for image transmission via multiple communication systems: FastAPI, ZeroMQ (ZMQ), ImageZMQ, grpc."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Timo Leitritz", email = "42964574+TimoIllusion@users.noreply.github.com" },
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "opencv-python>=4.13.0.92,<4.14.0",
  "fastapi>=0.136.0,<0.137.0",
  "uvicorn>=0.45.0,<0.46.0",
  "python-multipart>=0.0.26,<0.1.0",
  "torch>=2.0.0,<3.0.0",
  "torchvision",
  "yolov8>=0.0.2,<0.1.0",
  "requests>=2.33.1,<2.34.0",
  "loguru>=0.7.3,<0.8.0",
  "pyzmq>=27.1.0,<27.2.0",
  "imagezmq>=1.2.0,<1.3.0",
  "grpcio>=1.80.0,<1.81.0",
  "grpcio-tools>=1.80.0,<1.81.0",
  "ultralytics>=8.4.41,<8.5.0",
]

[project.optional-dependencies]
test = [
  "flake8>=6,<8",
  "numpy>=1.24,<3",
  "opencv-python-headless>=4.8,<5",
  "pillow>=10,<12",
  "pytest>=7,<8",
  "pytest-asyncio>=0.21,<1",
  "httpx>=0.27,<1",
]
dev = [
  "pre-commit>=3.7,<4",
  "ruff>=0.6,<1",
]

[project.urls]
Homepage = "https://github.com/TimoIllusion/inference-streaming-benchmark"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["inference_streaming_benchmark*"]

[tool.setuptools_scm]
# Version is derived from the latest git tag (e.g. v0.1.3 -> 0.1.3).
# Fallback used when the working tree has no tag at all (fresh clones of shallow CI checkouts).
fallback_version = "0.0.0"

[tool.ruff]
line-length = 128
target-version = "py310"
extend-exclude = [
  ".venv",
  "Ultralytics",
  "**/*_pb2.py",
  "**/*_pb2.pyi",
  "**/*_pb2_grpc.py",
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E203"]

[tool.ruff.lint.per-file-ignores]
"tests/**.py" = ["E402"]
"inference_streaming_benchmark/transports/http_multipart/transport.py" = ["B008"]

[tool.ruff.format]
line-ending = "auto"

